# openapi/admin.yaml
openapi: 3.0.3
info:
  title: CMS Admin API
  version: 0.2.0
  description: |
    CMS 管理后台接口契约,前后端唯一事实源。所有路径字面带 /api/admin 前缀(网关按此前缀转发)。

    修改流程:改本文件 → `pnpm gen:api` 双端生成 → 双端实现。
    生成物(apps/admin/src/api/generated/、apps/server/gen/)禁止手改。

    响应包装:传输层统一为 `{code, message, data, logID}`(code 等于 HTTP 状态码;
    logID 为本次请求的日志关联 ID,同时回传在 `X-Log-Id` 响应头,客户端可经
    `X-Request-Id` 请求头透传网关生成的 ID);错误响应为 `{code, message, logID}`。
    本契约描述的是 data 载荷,admin 在 mutator(src/api/client.ts)统一解包。
...
security:
  - bearerAuth: []
paths:
...
  /api/admin/roles:
    get:
      operationId: listRoles
      summary: 角色列表
      tags:
        - roles
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: keyword
          in: query
          schema:
            type: string
          description: 按角色名/编码模糊匹配
        - name: status
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: 角色分页列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleListResponse"
    post:
      operationId: createRole
      summary: 新建角色
      tags:
        - roles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleRequest"
      responses:
        "200":
          description: 新建后的角色
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleItem"
        "409":
          description: 角色编码已存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/admin/roles/{id}:
    get:
      operationId: getRole
      summary: 角色详情(含权限点 ID)
      tags:
        - roles
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: 角色详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleItem"
    put:
      operationId: updateRole
      summary: 编辑角色
      tags:
        - roles
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleRequest"
      responses:
        "200":
          description: 更新后的角色
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleItem"
        "403":
          description: 内置角色不可修改编码
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteRole
      summary: 删除角色(仍有用户绑定时拒绝)
      tags:
        - roles
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: 删除成功
        "403":
          description: 内置角色不可删除
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: 角色仍有用户绑定
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/admin/roles/{id}/permissions:
    put:
      operationId: updateRolePermissions
      summary: 角色分配权限(全量覆盖)
      tags:
        - roles
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PermissionIdsRequest"
      responses:
        "204":
          description: 分配成功
...
  /api/admin/uploads/images:
    post:
      operationId: initImageUpload
      summary: 初始化图片分片上传会话(权限码 media:image:upload)
      tags:
        - uploads
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitUploadRequest"
      responses:
        "200":
          description: 新建的上传会话
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadSession"
        "400":
          description: 文件类型不支持、大小非法或分组不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/admin/uploads/videos:
    post:
      operationId: initVideoUpload
      summary: 初始化视频分片上传会话(权限码 media:video:upload;视频上限 2GB)
      tags:
        - uploads
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitUploadRequest"
      responses:
        "200":
          description: 新建的上传会话
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadSession"
        "400":
          description: 文件类型不支持、大小非法或分组不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/admin/uploads/{uploadId}/chunks/{index}:
    put:
      operationId: uploadChunk
      summary: 上传分片(octet-stream;重复上传同一分片幂等覆盖;仅会话属主可用)
      tags:
        - uploads
      parameters:
        - name: uploadId
          in: path
          required: true
          schema:
            type: string
        - name: index
          in: path
          required: true
          schema:
            type: integer
            minimum: 0
          description: 分片索引,0 起
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "204":
          description: 分片已保存
        "400":
          description: 分片索引越界或单片超过 chunkSize
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: 会话不存在(含非属主访问,不区分)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/admin/uploads/{uploadId}:
    get:
      operationId: getUploadSession
      summary: 查询上传会话状态(断点续传用;仅会话属主可用)
      tags:
        - uploads
      parameters:
        - name: uploadId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 会话状态
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadSession"
        "404":
          description: 会话不存在(含非属主访问,不区分)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: abortUpload
      summary: 中止并清理上传会话(仅会话属主可用)
      tags:
        - uploads
      parameters:
        - name: uploadId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: 会话已清理
        "404":
          description: 会话不存在(含非属主访问,不区分)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/admin/uploads/{uploadId}/complete:
    post:
      operationId: completeUpload
      summary: 合并分片并走媒体上传管线(校验/存储/提取/落库/操作日志;仅会话属主可用)
      tags:
        - uploads
      parameters:
        - name: uploadId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 上传完成的媒体资源(精简字段)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadedMedia"
        "400":
          description: 分片不齐全或总大小不符
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: 会话不存在(含非属主访问,不区分)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: HTTP 状态码
        message:
          type: string
          description: 错误描述
...
    RoleItem:
      type: object
      required:
        - id
        - code
        - name
        - status
        - isBuiltin
        - permissionIds
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        name:
          type: string
        remark:
          type: string
        status:
          type: boolean
        isBuiltin:
          type: boolean
        permissionIds:
          type: array
          description: 角色拥有的权限点 ID
          items:
            type: integer
            format: int64
    RoleListResponse:
      type: object
      required:
        - list
        - total
      properties:
        list:
          type: array
          items:
            $ref: "#/components/schemas/RoleItem"
        total:
          type: integer
...
    RoleRequest:
      type: object
      required:
        - code
        - name
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 64
        name:
          type: string
          minLength: 1
          maxLength: 64
        remark:
          type: string
          maxLength: 255
        status:
          type: boolean
          default: true
    PermissionIdsRequest:
      type: object
      required:
        - permissionIds
      properties:
        permissionIds:
          type: array
          items:
            type: integer
            format: int64
...
    InitUploadRequest:
      type: object
      required:
        - fileName
        - size
      properties:
        fileName:
          type: string
          minLength: 1
          maxLength: 255
          description: 原始文件名(扩展名决定类型校验)
        size:
          type: integer
          format: int64
          minimum: 1
          description: 文件总字节数
        groupId:
          type: integer
          format: int64
          description: 可选,上传到的分组 ID;缺省或 0 为未分组
    UploadSession:
      type: object
      required:
        - uploadId
        - kind
        - fileName
        - size
        - chunkSize
        - chunkCount
        - uploadedIndexes
      properties:
        uploadId:
          type: string
          description: 会话 ID,后续分片/状态/合并/中止均以此定位
        kind:
          type: string
          enum: [image, video]
          description: 媒体类型
        fileName:
          type: string
        size:
          type: integer
          format: int64
          description: 文件总字节数
        chunkSize:
          type: integer
          format: int64
          description: 单片字节数(最后一片可小于该值)
        chunkCount:
          type: integer
          description: 分片总数
        uploadedIndexes:
          type: array
          description: 已落盘分片索引(0 起,升序),断点续传据此跳过
          items:
            type: integer
    UploadedMedia:
      type: object
      required:
        - id
        - kind
        - url
        - origName
        - size
      properties:
        id:
          type: integer
          format: int64
          description: 媒体资源 ID
        kind:
          type: string
          enum: [image, video]
        url:
          type: string
          description: 外网访问地址(CDN 直链);local 存储为空串,前端回退 /files/{fileId}/content
        origName:
          type: string
        size:
          type: integer
          format: int64
          description: 字节